Add the notion of Component
Components is basically a set of target on which we bind a user
option to enable or disable it.

There are several ways to create a component:

* Creating explicitely in the meson.build file:

```
component = component('componentname',
    targets: [],
    default: 'disabled', # disabled by default
    dependencies: [another_component], # @componentname is disabled
                                       # if @another_component is disabled
)
```

* Specifying it as a string or reusing previsouly creating one creating a target

```
target = library(..., components: ['componentname', component])
```

* Specifying it as a string or reusing previsouly creating one for all targets in a subdir

It means that all targets created inside those subdirectories will be contained in
the specified components. You can do so with:

```
subdir('somesubdir', components: 'componentname')
```